update: tambah pengajuan sertifikasi

+++ database/migrations/2024_02_20_105047_create_sertifikasis_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('nama_pemohon');
             $table->string('alamat');
             //sertifikasi benih untuk..
-            $table->integer('luas_pertanaman');
+            $table->float('luas_pertanaman');
             $table->foreignId('id_varietas');
             $table->date('tanggal_sebar');
             $table->date('tanggal_tanam');
